Refactor import of mock data
