+++ packages/lib-classifier/src/plugins/tasks/SurveyTask/mock-data/index.js
@@ -1,0 +1,1 @@
+export { default as task } from './task'
